requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
apache-airflow==2.8.0
apache-airflow-providers-databricks==4.9.0
apache-airflow-providers-microsoft-azure==7.0.0
pyodbc==5.0.1
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
confluent-kafka==2.3.0
azure-storage-blob==12.19.0
